[AIRFLOW-6160] Move _check_and_change_state_before_execution to LocalTaskJob
